FIX: nighly build on github also for SDL2 64bit
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@9185 8aca7d54-2c30-db11-9de9-000461428c89

+++ gui/schedule_gui.cc
@@ -601,7 +601,7 @@ DBG_MESSAGE("schedule_gui_t::action_triggered()","comp=%p combo=%p",comp,&line_s
 		// update line schedule via tool!
 		tool_t *tool = create_tool( TOOL_CHANGE_LINE | SIMPLE_TOOL );
 		cbuffer_t buf;
-		buf.printf( "c,0,%i,%ld,", (int)schedule->get_type(), (long)old_schedule );
+		buf.printf( "c,0,%i,%ld,", (int)schedule->get_type(), (long)(intptr_t)old_schedule );
 		schedule->sprintf_schedule( buf );
 		tool->set_default_param(buf);
 		welt->set_tool( tool, player );
